+++ R/all-selectors.r
@@ -1,0 +1,22 @@
+#' Names of different selectors
+#'
+#' The package makes available character vectors that can be used to extract
+#' spectral data for specific subsets of the filters, screens, etc. Some
+#' correspond to supplier names, other to materials, types, colours and typical
+#' uses. The data objects described here list the names of all these vectors
+#' of names corresponding to the different collections of spectra..
+#'
+#' @docType data
+#' @keywords datasets
+#' @format A vector of character strings.
+#'
+#' @examples
+#' all_selectors
+#'
+#' @rdname all_selectors
+#' @aliases selectors
+#'
+"all_filter_selectors"
+
+#' @rdname all_selectors
+"all_screen_selectors"