+++ R/screens.R
@@ -1,0 +1,48 @@
+#' Spectral data for filters made from different materials
+#'
+#' Names of datasets containing the wavelengths and tabulated values for
+#' fractional spectral transmittance grouped by supplier.
+#'
+#' @docType data
+#' @keywords datasets
+#' @format A vector of character strings.
+#'
+#' @examples
+#' arrigoni_screens
+#'
+#' # select filters supplied by Arrigoni
+#' screens.mspct[arrigoni_screens[1:5]]
+#'
+#' @seealso \code{\link{screens.mspct}}
+#'
+#' @rdname screen_suppliers
+#'
+"arrigoni_screens"
+
+#' @rdname screen_suppliers
+#'
+"criadolopez_screens"
+
+#' @rdname screen_suppliers
+#'
+"howitec_screens"
+
+#' @rdname screen_suppliers
+#'
+"huachangyarns_screens"
+
+#' @rdname screen_suppliers
+#'
+"jiangsuhuachangyarns_screens"
+
+#' @rdname screen_suppliers
+#'
+"mallastextiles_screens"
+
+#' @rdname screen_suppliers
+#'
+"oerlemansplastics_screens"
+
+#' @rdname screen_suppliers
+#'
+"svensson_screens"