+++ data-raw/save-filter-data.r
@@ -26,7 +26,11 @@ filters.mspct <-
   xl_horticulture.mspct, zeiss.mspct, glass_windows.mspct, photography_filters.mspct,
   kolarivision.mspct)
 
+filters.mspct <- thin_wl(filters.mspct, max.wl.step = 5, max.slope.delta = 0.00025)
+
 filters.mspct <- filters.mspct[sort(names(filters.mspct))]
+names(filters.mspct)
+
 midopt_filters <- midopt
 lee_gels <- lee_filters <- lee
 rosco_gels <- rosco_filters <- rosco
@@ -48,10 +52,11 @@ zeiss_filters <- grep("Zeiss", photography_filters, value = TRUE, ignore.case = 
 baader_filters <- grep("Baader", photography_filters, value = TRUE, ignore.case = TRUE)
 tangsinuo_filters <- grep("Tangsinuo", photography_filters, value = TRUE, ignore.case = TRUE)
 uvroptics_filters <- grep("StraightEdgeU", photography_filters, value = TRUE, ignore.case = TRUE)
-fake_unbranded_filters <- grep("Knight|fake", photography_filters, value = TRUE, ignore.case = TRUE)
+knightx_filters <- grep("Knight", photography_filters, value = TRUE, ignore.case = TRUE)
+fake_unbranded_filters <- grep("unbranded|fake", photography_filters, value = TRUE, ignore.case = TRUE)
 photography_filters <-
-  sort(c(photography_filters, zeiss_filters, kolarivision_filters,
-         fotga_filters, tangsinuo_filters, purshee_filters))
+  unique(sort(c(photography_filters, zeiss_filters, kolarivision_filters,
+         fotga_filters, tangsinuo_filters, purshee_filters)))
 courtaulds_filters <- courtaulds
 xl_horticulture_filters <- xl_horticulture
 bpi_visqueen_filters <- bpi_visqueen
@@ -91,9 +96,9 @@ long_pass_filters <- c(grep("GG|WG|OG|RG|LP", names(filters.mspct), value = TRUE
 short_pass_filters <- grep("KG|SP", names(filters.mspct), value = TRUE)
 
 # all_accessors <- sort(setdiff(ls(pattern = "*"), ls(pattern = "*.mspct")))
-all_filter_accessors <- sort(grep("_filters$", ls(pattern = "*"), value = TRUE))
+all_filter_selectors <- sort(grep("_filters$", ls(pattern = "*"), value = TRUE))
 
-save(filters.mspct, all_filter_accessors,
+save(filters.mspct, all_filter_selectors,
      bpi_visqueen_filters, courtaulds_filters, etola_filters, evonik_filters, foiltek_filters,
      lee_filters, lee_gels, mcdermit_filters, petri_dishes, plexiglas_filters,
      rosco_filters, rosco_gels, schott_filters, xl_horticulture_filters, midopt_filters,
@@ -102,7 +107,8 @@ save(filters.mspct, all_filter_accessors,
      polystyrene_filters, polyester_filters, polyvynil_chloride_filters,
      photography_filters, hoya_filters, firecrest_filters, bw_filters, zomei_filters, fotga_filters,
      haida_filters, kenko_filters, tiffen_filters, baader_filters, uvroptics_filters, tangsinuo_filters,
-     heliopan_filters, rocolax_filters, zeiss_filters, kenko_filters, purshee_filters, fake_unbranded_filters,
+     heliopan_filters, rocolax_filters, zeiss_filters, kenko_filters, purshee_filters,
+     knightx_filters, fake_unbranded_filters,
      uvir_cut_filters, theatrical_gels, optical_glass_filters,
      plastic_films, plastic_sheets, plastic_film_filters, plastic_sheet_filters,
      clear_filters, yellow_filters, orange_filters, green_filters, blue_green_filters,